Implemented functionality to update specific game fields without full validation
This change introduces an `updateGameField` method to `GameRepository` and its implementation. This allows for updating specific fields of a game document in Firestore without requiring a full game object comparison, which can be useful in scenarios like updating the game state to "GAME_ABORTED" when a user leaves during the "CHECK_READY" phase.

Additionally, a minor correction was made to the `winnerId` type in `GameDto` from `Int?` to `String?`.