Implemented "You Won" screen when opponent leaves the game.
The `LeaveGameUseCase` now sets the `gameState` to `USER_LEFT` when a player leaves an ongoing game.
A new `USER_LEFT` state has been added to the `GameState` enum.
The `GameScreen` now displays a "You have won!!" message if the `gameState` is `USER_LEFT` and the current user is the winner.