Removed JoinGameUseCase and updated Game logic
Deleted the `JoinGameUseCase`, `playerJoined` fields as its functionality is not needed. It was intended to be a intermediate state to wait for players to join the Game before the Room was deleted. Once both players had joined the game, GameState transitioned from 'WAITING_FOR_PLAYERS' to 'CHECK_READY' and the room was deleted from firestore through the CloseRoomUseCase.

- Deleted JoinGameUseCase
- Deleted player1Joined and player2Joined from Game and GameDto.
- Updated `GameState` to remove `WAITING_FOR_PLAYERS`, as `CHECK_READY` now covers this.
- Simplified `GameViewModel` by removing `JoinGameUseCase` and related logic.
- Updated `WaitRoomUseCase` to set the initial game state to `CHECK_READY`.
- Modified `GameScreen` to only check for `CHECK_READY` state for the ready check section.